:recycle: refactor a little
